requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.3.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
Pillow==9.4.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
